adding storage limit for building in progress also need to divide entities and dtos

+++ src/main/java/com/creazione/space_learning/game/Item.java
@@ -9,9 +9,9 @@ import lombok.*;
 @AllArgsConstructor
 public class Item {
     private ResourceType name;
-    private double quantity;
+    private long quantity;
 
-    public void addQuantity(double quantity) {
+    public void addQuantity(long quantity) {
         this.quantity += quantity;
     }
 }
